services/quota_mgmt/pyproject.toml
[tool.poetry]
name = "quota-mgmt"
version = "0.1.0"
description = ""
authors = ["akirawuc <stu970029@gmail.com>"]
readme = "README.md"
packages = [{include = "quota_mgmt"}]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-firestore = "^2.11.1"
functions-framework = "^3.4.0"
flask = "^2.3.2"
flask-cors = "^4.0.0"
web3 = "^6.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
